removed old crf and cv pipeline (now merged into one)
